requirements.txt
PyYAML==6.0.2
numpy==2.2.4
opencv-python==4.12.0.88
